MDEV-28310 Missing binlog data for INSERT .. ON DUPLICATE KEY UPDATE
MDEV-17614 fixed replication unsafety for INSERT ON DUP KEY UPDATE on two or more unique non-autoincrement key table leaving a flaw for a case
of the INSERT containing multiple records and the duplicate errored out one
is not the last.
Combined with MIXED @@binlog_format that could lead to loss of replication
event in binlog.

This case has been already fixed in 10.5.2 by 91ab42a823.
That commit actually not merely optimized `THD::decide_logging_format_low()`
via relocation but also corrected its execution logic that had assumed
`table->write_set` is stable throughout records processing loop
but which in fact is not.